Working on getting SPI functional on RPi.
