+++ Science/Science Onboard/RoverMain.cs
@@ -8,6 +8,7 @@ using Scarlet.Components.Motors;
 using Scarlet.Components.Sensors;
 using Scarlet.IO;
 using Scarlet.IO.BeagleBone;
+using Scarlet.IO.RaspberryPi;
 using Scarlet.Science;
 using Scarlet.Utilities;
 
@@ -19,8 +20,7 @@ namespace Science
         private static string IP = Constants.DEFAULT_SERVER_IP;
         private static int PortTCP = Constants.DEFAULT_PORT_TCP;
         private static int PortUDP = Constants.DEFAULT_PORT_UDP;
-        private static BBBPinManager.ApplicationMode ApplyDevTree = BBBPinManager.ApplicationMode.APPLY_IF_NONE;
-        private static IDigitalIn IntTestIn;
+        public static BBBPinManager.ApplicationMode ApplyDevTree = BBBPinManager.ApplicationMode.APPLY_IF_NONE;
 
         static void Main(string[] Args)
 		{
@@ -33,9 +33,11 @@ namespace Science
             Log.Begin();
             Log.ForceOutput(Log.Severity.INFO, Log.Source.OTHER, "Science Station - Rover Side");
 
-            BeagleBone.Initialize(SystemMode.DEFAULT, true);
+            //BeagleBone.Initialize(SystemMode.DEFAULT, true);
             Log.SetSingleOutputLevel(Log.Source.HARDWAREIO, Log.Severity.DEBUG);
-            TestPWM();
+
+            RaspberryPi.Initialize();
+            RPiTests.TestSPI();
 
             IOHandler = new IOHandler();
             Client.Start(IP, PortTCP, PortUDP, Constants.CLIENT_NAME);
@@ -53,176 +55,6 @@ namespace Science
             Console.ReadKey();
             Environment.Exit(0);
 		}
-
-        private static void TestDigO()
-        {
-            BBBPinManager.AddMappingGPIO(BBBPin.P8_08, true, Scarlet.IO.ResistorState.PULL_DOWN);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-            IDigitalOut Output = new DigitalOutBBB(BBBPin.P8_08);
-            bool Value = false;
-            for (int i = 0; i < 50; i++)
-            {
-                Output.SetOutput(Value);
-                Value = !Value;
-                Thread.Sleep(100);
-            }
-            Output.SetOutput(false);
-        }
-
-        private static void TestDigI()
-        {
-            BBBPinManager.AddMappingGPIO(BBBPin.P9_12, false, Scarlet.IO.ResistorState.PULL_DOWN);
-            //if (ApplyDevTree) { BBBPinManager.ApplyPinSettings(); }
-            IDigitalIn Input = new DigitalInBBB(BBBPin.P9_12);
-            for(int i = 0; i < 50; i++)
-            {
-                Log.Output(Log.Severity.DEBUG, Log.Source.HARDWAREIO, "Input is " + Input.GetInput());
-                Thread.Sleep(250);
-            }
-        }
-
-        private static void TestPWM()
-        {
-            BBBPinManager.AddMappingPWM(BBBPin.P9_14);
-            BBBPinManager.AddMappingPWM(BBBPin.P9_16);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-            IPWMOutput OutA = PWMBBB.PWMDevice1.OutputA;
-            IPWMOutput OutB = PWMBBB.PWMDevice1.OutputB;
-            PWMBBB.PWMDevice1.SetFrequency(5000);
-            OutA.SetEnabled(true);
-            OutB.SetEnabled(true);
-            int Cycle = 0;
-            while (true)
-            {
-                float A = (float)((Math.Sin(Cycle * Math.PI / 180.000D) + 1) / 2); // Sine waves! Fun!
-                float B = (float)((Math.Sin(Cycle * Math.PI / 360.000D) + 1) / 2);
-
-                OutA.SetOutput(A);
-                OutB.SetOutput(B);
-
-                Thread.Sleep(50);
-                Cycle += 20;
-            }
-        }
-
-        private static void TestMotor()
-        {
-            BBBPinManager.AddMappingPWM(BBBPin.P9_14);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-            IPWMOutput MotorOut = PWMBBB.PWMDevice1.OutputA;
-            MotorOut.SetEnabled(true);
-            TalonMC Motor = new TalonMC(MotorOut, 0.2F);
-            Log.SetSingleOutputLevel(Log.Source.MOTORS, Log.Severity.DEBUG);
-            Motor.RampUp = 0.5F;
-            //Motor.Speed = 0.2F;
-            Motor.UpdateState();
-            while (true)
-            {
-                Log.Output(Log.Severity.DEBUG, Log.Source.MOTORS, "Outputs: " + Motor.Speed + ", " + ((PWMOutputBBB)MotorOut).GetOutput() + ", " + ((PWMOutputBBB)MotorOut).GetFrequency());
-                //Motor.UpdateState();
-                Thread.Sleep(100);
-            }
-            /*int Cycle = 0;
-            while (true)
-            {
-                Motor.UpdateState();
-                float Spd = (float)((Math.Sin(Cycle * Math.PI / 360.000D) + 1) / 40) + 0.1F;
-                Log.Output(Log.Severity.DEBUG, Log.Source.MOTORS, "Outputting " + Spd + ", currently " + Motor.Speed);
-                Motor.Speed = Spd;
-                Thread.Sleep(100);
-                Cycle += 15;
-            }*/
-        }
-
-        private static void TestPWMLow()
-        {
-            BBBPinManager.AddMappingPWM(BBBPin.P9_14);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-
-            PWMPortMM Port = new PWMPortMM(PWMPortEnum.PWM1_A);
-            Port.PeriodNS = 1000000;
-            Port.DutyPercent = 0;
-            Port.RunState = true;
-            while (true)
-            {
-                for (int i = 0; i < 100; i++)
-                {
-                    Port.DutyPercent = i;
-                    Thread.Sleep(10);
-                }
-                Port.DutyPercent = 100;
-                Thread.Sleep(50);
-                for (int i = 100; i > 0; i--)
-                {
-                    Port.DutyPercent = i;
-                    Thread.Sleep(10);
-                }
-                Port.DutyPercent = 0;
-                Thread.Sleep(50);
-            }
-        }
-        
-        private static void TestI2C()
-        {
-            BBBPinManager.AddMappingGPIO(BBBPin.P8_08, true, Scarlet.IO.ResistorState.PULL_DOWN);
-            BBBPinManager.AddMappingsI2C(BBBPin.P9_24, BBBPin.P9_26);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-            VEML6070 UV = new VEML6070(I2CBBB.I2CBus1);
-            Log.SetSingleOutputLevel(Log.Source.SENSORS, Log.Severity.DEBUG);
-            for (int i = 0; i < 20; i++)
-            {
-                UV.UpdateState();
-                Log.Output(Log.Severity.DEBUG, Log.Source.SENSORS, "UV Reading: " + UV.GetData());
-                Thread.Sleep(200);
-            }
-        }
-
-        private static void TestSPI()
-        {
-            BBBPinManager.AddMappingsSPI(BBBPin.P9_21, BBBPin.NONE, BBBPin.P9_22);
-            BBBPinManager.AddMappingSPI_CS(BBBPin.P9_12);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-            IDigitalOut CS_Thermo = new DigitalOutBBB(BBBPin.P9_12);
-            MAX31855 Thermo = new MAX31855(SPIBBB.SPIBus0, CS_Thermo);
-            Log.SetSingleOutputLevel(Log.Source.SENSORS, Log.Severity.DEBUG);
-            for (int i = 0; i < 100; i++)
-            {
-                Thermo.UpdateState();
-                Log.Output(Log.Severity.DEBUG, Log.Source.SENSORS, "Thermocouple Data, Faults: " + string.Format("{0:G}", Thermo.GetFaults()) + ", Internal: " + Thermo.GetInternalTemp() + ", External: " + Thermo.GetExternalTemp() + " (Raw: " + Thermo.GetRawData() + ")");
-                Thread.Sleep(500);
-            }
-        }
-
-        private static void TestADC()
-        {
-            BBBPinManager.AddMappingADC(BBBPin.P9_36);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-            IAnalogueIn Input = new AnalogueInBBB(BBBPin.P9_36);
-            for(int i = 0; i < 200; i++)
-            {
-                Log.Output(Log.Severity.DEBUG, Log.Source.HARDWAREIO, "ADC Input: " + Input.GetInput() + " (Raw: " + Input.GetRawInput() + ")");
-                Thread.Sleep(100);
-            }
-        }
-
-        private static void TestInterrupt()
-        {
-            BBBPinManager.AddMappingGPIO(BBBPin.P9_12, true, Scarlet.IO.ResistorState.PULL_DOWN);
-            BBBPinManager.ApplyPinSettings(ApplyDevTree);
-            IntTestIn = new DigitalInBBB(BBBPin.P9_12);
-            IntTestIn.RegisterInterruptHandler(GetInterrupt, InterruptType.ANY_EDGE);
-            Log.Output(Log.Severity.DEBUG, Log.Source.HARDWAREIO, "Interrupt handler added.");
-            while(true)
-            {
-                //Log.Output(Log.Severity.DEBUG, Log.Source.HARDWAREIO, "State: " + IntTestIn.GetInput());
-                Thread.Sleep(100);
-            }
-        }
-
-        public static void GetInterrupt(object Senser, InputInterrupt Event)
-        {
-            Log.Output(Log.Severity.DEBUG, Log.Source.HARDWAREIO, "Interrupt Received! Now " + Event.NewState);
-        }
 
         private static void ParseArgs(string[] Args)
         {
