+++ Shared/Scarlet/IO/RaspberryPi/RaspberryPi.cs
@@ -6,10 +6,7 @@ namespace Scarlet.IO.RaspberryPi
 {
     public static class RaspberryPi
     {
-        private const string MAIN_LIB = "libwiringPi.so";
-        private const string I2C_LIB = "libwiringPiI2C.so";
-        private const string SPI_LIB = "libwiringPiSPI.so";
-        private const string UART_LIB = "libwiringPiUART.so";
+        private const string WIRING_PI_LIB = "libWiringPi-2.44.so";
 
         private static bool P_Initialized = false;
         internal static bool Initialized
@@ -18,9 +15,8 @@ namespace Scarlet.IO.RaspberryPi
             private set { P_Initialized = value; }
         }
 
-
         /// <returns>Always 0, You can ignore this.</returns>
-        [DllImport(MAIN_LIB, EntryPoint = "wiringPiSetupPhys")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "wiringPiSetupPhys")]
         private static extern int Ext_SetupGPIO();
 
         /// <summary>
@@ -42,7 +38,7 @@ namespace Scarlet.IO.RaspberryPi
 
         #region GPIO
 
-        [DllImport(MAIN_LIB, EntryPoint = "pinMode")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "pinMode")]
         private static extern void Ext_SetPinMode(int Pin, int Mode);
 
         internal static void SetPinMode(int Pin, PinMode Mode)
@@ -51,7 +47,7 @@ namespace Scarlet.IO.RaspberryPi
             Ext_SetPinMode(Pin, (int)Mode);
         }
 
-        [DllImport(MAIN_LIB, EntryPoint = "pullUpDnControl")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "pullUpDnControl")]
         private static extern void Ext_SetResistor(int Pin, int ResMode);
 
         internal static void SetResistor(int Pin, ResistorState ResMode)
@@ -63,7 +59,7 @@ namespace Scarlet.IO.RaspberryPi
             Ext_SetResistor(Pin, ResistorVal);
         }
 
-        [DllImport(MAIN_LIB, EntryPoint = "digitalRead")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "digitalRead")]
         private static extern int Ext_DigitalRead(int Pin);
 
         internal static bool DigitalRead(int Pin)
@@ -73,7 +69,7 @@ namespace Scarlet.IO.RaspberryPi
         }
 
 
-        [DllImport(MAIN_LIB, EntryPoint = "digitalWrite")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "digitalWrite")]
         private static extern void Ext_DigitalWrite(int Pin, int Value);
 
         internal static void DigitalWrite(int Pin, bool OutputVal)
@@ -88,7 +84,7 @@ namespace Scarlet.IO.RaspberryPi
 
         internal delegate void InterruptCallback();
 
-        [DllImport(MAIN_LIB, EntryPoint = "wiringPiISR")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "wiringPiISR")]
         private static extern int Ext_AddInterrupt(int Pin, int InterruptType, InterruptCallback Delegate);
 
         internal static void AddInterrupt(int Pin, int InterruptType, InterruptCallback Delegate)
@@ -101,7 +97,7 @@ namespace Scarlet.IO.RaspberryPi
 
         #region I2C
 
-        [DllImport(I2C_LIB, EntryPoint = "wiringPiI2CSetup")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "wiringPiI2CSetup")]
         private static extern int Ext_I2CSetup(int DeviceID);
 
         internal static void I2CSetup(byte DeviceID)
@@ -110,7 +106,7 @@ namespace Scarlet.IO.RaspberryPi
             if (SetupVal == -1) { throw new Exception("Unable to setup I2C device: " + DeviceID); }
         }
 
-        [DllImport(I2C_LIB, EntryPoint = "wiringPiI2CRead")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "wiringPiI2CRead")]
         private static extern int Ext_I2CRead(int DeviceID);
 
         internal static byte I2CRead(byte DeviceID)
@@ -119,7 +115,7 @@ namespace Scarlet.IO.RaspberryPi
             return (byte)Data;
         }
 
-        [DllImport(I2C_LIB, EntryPoint = "wiringPiI2CWrite")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "wiringPiI2CWrite")]
         private static extern int Ext_I2CWrite(int DeviceID, int Data);
 
         internal static void I2CWrite(byte DeviceID, byte Data)
@@ -131,12 +127,12 @@ namespace Scarlet.IO.RaspberryPi
 
         #region SPI
 
-        [DllImport(SPI_LIB, EntryPoint = "wiringPiSPISetup")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "wiringPiSPISetup")]
         private static extern void Ext_SPISetup(int BusNum, int Speed);
 
         internal static void SPISetup(int BusNum, int Speed) { Ext_SPISetup(BusNum, Speed); }
 
-        [DllImport(SPI_LIB, EntryPoint = "wiringPiSPIDataRW")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "wiringPiSPIDataRW")]
         private static extern int Ext_SPIRW(int BusNum, [In,Out] byte[] Data, int Length);
 
         internal static byte[] SPIRW(int BusNum, byte[] Data, int Length)
@@ -151,45 +147,44 @@ namespace Scarlet.IO.RaspberryPi
 
         #region UART
 
-        [DllImport(UART_LIB, EntryPoint = "serialOpen")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "serialOpen")]
         private static extern int Ext_SerialOpen(byte Device, int Baud);
 
         // Returns Device ID, -1 on error
         internal static int SerialOpen(byte Device, int Baud) { return Ext_SerialOpen(Device, Baud); }
 
-        [DllImport(UART_LIB, EntryPoint = "serialClose")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "serialClose")]
         private static extern void Ext_SerialClose(int DeviceID);
 
         internal static void SerialClose(int DeviceID) { Ext_SerialClose(DeviceID); }
 
-        [DllImport(UART_LIB, EntryPoint = "serialPuts")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "serialPuts")]
         private static extern void Ext_SerialPut(int DeviceID, [In,Out] byte[] Data);
 
         internal static void SerialPut(int DeviceID, byte[] Data) { Ext_SerialPut(DeviceID, Data); }
 
-        [DllImport(UART_LIB, EntryPoint = "serialDataAvail")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "serialDataAvail")]
         private static extern int Ext_SerialDataAvailable(int DeviceID);
 
         // Returns # bytes available to read
         internal static int SerialDataAvailable(int DeviceID) { return Ext_SerialDataAvailable(DeviceID); }
 
-        [DllImport(UART_LIB, EntryPoint = "serialFlush")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "serialFlush")]
         private static extern void Ext_SerialFlush(int DeviceID);
 
         internal static void SerialFlush(int DeviceID) { Ext_SerialFlush(DeviceID); }
 
-        [DllImport(UART_LIB, EntryPoint = "serialGetchar")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "serialGetchar")]
         private static extern byte Ext_SerialGetChar(int DeviceID);
 
         internal static byte SerialGetChar(int DeviceID) { return Ext_SerialGetChar(DeviceID); }
 
-        [DllImport(UART_LIB, EntryPoint = "serialPrintf")]
+        [DllImport(WIRING_PI_LIB, EntryPoint = "serialPrintf")]
         private static extern void Ext_SerialPrintf(int DeviceID, [In,Out] byte[] Data);
 
         internal static void SerialPrintf(int DeviceID, byte[] Data) { Ext_SerialPrintf(DeviceID, Data); }
 
         #endregion
-    
 
     }
 }
